Battle_Omok_AI/requirements.txt
pygame
pyyaml
pytest
